Change the type of AV1_COMP::bytes to uint64_t
In commit 699cefe, the type of AV1_PRIMARY::total_bytes was changed to
uint64_t. For consistency, the type of the related AV1_COMP::bytes
should also be changed to to uint64_t.

+++ av1/av1_cx_iface.c
@@ -6,7 +6,7 @@
  * was not distributed with this source code in the LICENSE file, you can
  * obtain it at www.aomedia.org/license/software. If the Alliance for Open
  * Media Patent License 1.0 was not distributed with this source code in the
- * PATENTS file, you can obtain it at www.aomedia.org/license/patent.
+ * PATENTS file, you can obtain it at www.aomedia.org/license/patent..
  */
 #include <limits.h>
 #include <stdint.h>
@@ -3332,7 +3332,7 @@ static aom_codec_err_t encoder_encode(aom_codec_alg_priv_t *ctx,
       if (ppi->cpi->oxcf.pass != 1) {
         ppi->total_time_compress_data += cpi->time_compress_data;
         ppi->total_recode_hits += cpi->frame_recode_hits;
-        ppi->total_bytes += (uint64_t)cpi->bytes;
+        ppi->total_bytes += cpi->bytes;
         for (int i = 0; i < MAX_MODES; i++) {
           ppi->total_mode_chosen_counts[i] += cpi->mode_chosen_counts[i];
         }
